Remove unused parameters
chartZoomP and chartZoomF are never used, and redundant, as both charts are synced an hence use the same zoomP and zoomF parameters.

Next step would be nice to get rid of the public variables and store them in a widget state, spanning both chart widgets from one. But that's up to you guys.